Use full artifact version during verification
This solves a problem when transitive dependencies ends up with the
same name. Before this commit the following dependencies resulted in a
verifcation error:

Main dep:
 * foo-1.0.0.jar

Transitive deps:
 * foo-platform-1.0.0-linux.jar
 * foo-platform-1.0.0-osx.jar
 * foo-platform-1.0.0-windows.jar

In this case the dependencyVerification list had three entries
containing the name 'foo-platform' with different checksums. This
resulted in a conflict during verification.

This commit makes use of all the info that follows the name to find the
exact dependency, e.g. '1.0.0-linux' in this example.